requirements.txt
# requirements.txt - VERSIONES ESTABLES Y COMPATIBLES
streamlit==1.31.0
google-generativeai==0.7.2
requests==2.31.0
PyPDF2==3.0.1
sentence-transformers==2.3.1
chromadb==0.4.22
numpy==1.24.3
pandas==2.0.3
gspread==5.12.4
google-auth==2.23.4